+++ robotcode/language_server/robotframework/parts/semantic_tokens.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import ast
 import asyncio
-import itertools
 import operator
 import re
 from dataclasses import dataclass
@@ -14,7 +13,7 @@ from typing import (
     AsyncGenerator,
     Dict,
     FrozenSet,
-    Generator,
+    List,
     Optional,
     Set,
     Tuple,
@@ -23,6 +22,7 @@ from typing import (
 )
 
 from ....utils.async_event import CancelationToken
+from ....utils.async_itertools import async_dropwhile, async_takewhile
 from ....utils.logging import LoggingDescriptor
 from ...common.language import language_id
 from ...common.lsp_types import (
@@ -35,11 +35,18 @@ from ...common.lsp_types import (
     SemanticTokenTypes,
 )
 from ...common.text_document import TextDocument
-from ..diagnostics.library_doc import BUILTIN_LIBRARY_NAME, KeywordMatcher, LibraryDoc
-from ..diagnostics.namespace import KeywordFinder, Namespace
+from ..diagnostics.library_doc import (
+    ALL_RUN_KEYWORDS_MATCHERS,
+    BUILTIN_LIBRARY_NAME,
+    KeywordDoc,
+    KeywordMatcher,
+    LibraryDoc,
+)
+from ..diagnostics.namespace import Namespace
 from ..utils.ast import (
     HasTokens,
     Token,
+    is_not_variable_token,
     iter_nodes,
     token_in_range,
     tokenize_variables,
@@ -50,6 +57,8 @@ if TYPE_CHECKING:
     from ..protocol import RobotLanguageServerProtocol
 
 from .protocol_part import RobotLanguageServerProtocolPart
+
+ROBOT_KEYWORD_INNER = "KEYWORD_INNER"
 
 
 class RobotSemTokenTypes(Enum):
@@ -68,6 +77,7 @@ class RobotSemTokenTypes(Enum):
     ARGUMENT = "argument"
     VARIABLE = "variable"
     KEYWORD = "keywordCall"
+    KEYWORD_INNER = "keywordCallInner"
     NAME = "nameCall"
     CONTINUATION = "continuation"
     SEPARATOR = "separator"
@@ -118,7 +128,7 @@ class RobotSemanticTokenProtocolPart(RobotLanguageServerProtocolPart):
         parent.semantic_tokens.token_modifiers += [e for e in RobotSemTokenModifiers]
 
         parent.semantic_tokens.collect_full.add(self.collect_full)
-        parent.semantic_tokens.collect_range.add(self.collect_range)
+        # parent.semantic_tokens.collect_range.add(self.collect_range)
         # parent.semantic_tokens.collect_full_delta.add(self.collect_full_delta)
 
     @classmethod
@@ -161,6 +171,7 @@ class RobotSemanticTokenProtocolPart(RobotLanguageServerProtocolPart):
             frozenset({RobotToken.ARGUMENT}): (RobotSemTokenTypes.ARGUMENT, None),
             frozenset({RobotToken.VARIABLE, RobotToken.ASSIGN}): (RobotSemTokenTypes.VARIABLE, None),
             frozenset({RobotToken.KEYWORD}): (RobotSemTokenTypes.KEYWORD, None),
+            frozenset({ROBOT_KEYWORD_INNER}): (RobotSemTokenTypes.KEYWORD_INNER, None),
             frozenset({RobotToken.NAME}): (RobotSemTokenTypes.NAME, None),
             frozenset({RobotToken.CONTINUATION}): (RobotSemTokenTypes.CONTINUATION, None),
             frozenset({RobotToken.SEPARATOR}): (RobotSemTokenTypes.SEPARATOR, None),
@@ -192,7 +203,6 @@ class RobotSemanticTokenProtocolPart(RobotLanguageServerProtocolPart):
     async def generate_sem_sub_tokens(
         cls,
         namespace: Namespace,
-        finder: KeywordFinder,
         builtin_library_doc: Optional[LibraryDoc],
         token: Token,
         node: ast.AST,
@@ -284,7 +294,7 @@ class RobotSemanticTokenProtocolPart(RobotLanguageServerProtocolPart):
                         SemanticTokenTypes.OPERATOR,
                     )
                 if builtin_library_doc is not None and kw in builtin_library_doc.keywords:
-                    doc = await finder.find_keyword(token.value)
+                    doc = await namespace.find_keyword(token.value)
                     if (
                         doc is not None
                         and doc.libname == cls.BUILTIN_MATCHER
@@ -305,7 +315,6 @@ class RobotSemanticTokenProtocolPart(RobotLanguageServerProtocolPart):
         token: Token,
         node: ast.AST,
         namespace: Namespace,
-        finder: KeywordFinder,
         builtin_library_doc: Optional[LibraryDoc],
     ) -> AsyncGenerator[SemTokenInfo, None]:
         from robot.parsing.lexer.tokens import Token as RobotToken
@@ -315,36 +324,242 @@ class RobotSemanticTokenProtocolPart(RobotLanguageServerProtocolPart):
             for sub_token in tokenize_variables(
                 token, ignore_errors=True, identifiers="$" if token.type == RobotToken.KEYWORD_NAME else "$@&%"
             ):
-                async for e in self.generate_sem_sub_tokens(namespace, finder, builtin_library_doc, sub_token, node):
+                async for e in self.generate_sem_sub_tokens(namespace, builtin_library_doc, sub_token, node):
                     yield e
 
         else:
-            async for e in self.generate_sem_sub_tokens(namespace, finder, builtin_library_doc, token, node):
+            async for e in self.generate_sem_sub_tokens(namespace, builtin_library_doc, token, node):
                 yield e
+
+    async def generate_run_kw_tokens(
+        self,
+        namespace: Namespace,
+        builtin_library_doc: Optional[LibraryDoc],
+        kw_doc: Optional[KeywordDoc],
+        kw_token: Token,
+        arguments: List[Token],
+        node: ast.AST,
+    ) -> AsyncGenerator[Tuple[Token, ast.AST], None]:
+        from robot.parsing.lexer import Token as RobotToken
+        from robot.utils.escaping import unescape
+
+        async def skip_non_data_tokens() -> AsyncGenerator[Tuple[Token, ast.AST], None]:
+            nonlocal arguments
+            while arguments and arguments[0] and arguments[0].type in RobotToken.NON_DATA_TOKENS:
+
+                yield arguments[0], node,
+                arguments = arguments[1:]
+
+        yield kw_token, node
+
+        if kw_doc is not None and kw_doc.is_any_run_keyword():
+            async for b in skip_non_data_tokens():
+                yield b
+
+            if kw_doc.is_run_keyword() and len(arguments) > 0:
+                token = arguments[0]
+                async for b in self.generate_run_kw_tokens(
+                    namespace,
+                    builtin_library_doc,
+                    await namespace.find_keyword(unescape(token.value)) if is_not_variable_token(token) else None,
+                    RobotToken(ROBOT_KEYWORD_INNER, token.value, token.lineno, token.col_offset, token.error),
+                    arguments[1:],
+                    node,
+                ):
+                    yield b
+            elif kw_doc.is_run_keyword_with_condition() and len(arguments) > 0:
+                yield arguments[0], node,
+                arguments = arguments[1:]
+
+                async for b in skip_non_data_tokens():
+                    yield b
+
+                if len(arguments) > 0:
+                    token = arguments[0]
+                    async for b in self.generate_run_kw_tokens(
+                        namespace,
+                        builtin_library_doc,
+                        await namespace.find_keyword(unescape(token.value)) if is_not_variable_token(token) else None,
+                        RobotToken(ROBOT_KEYWORD_INNER, token.value, token.lineno, token.col_offset, token.error),
+                        arguments[1:],
+                        node,
+                    ):
+                        yield b
+            elif kw_doc.is_run_keywords() and len(arguments) > 0:
+                has_separator = False
+                while arguments:
+                    async for b in skip_non_data_tokens():
+                        yield b
+
+                    if not arguments:
+                        break
+
+                    token = arguments[0]
+                    arguments = arguments[1:]
+
+                    if token.value == "AND":
+                        yield RobotToken(
+                            RobotToken.ELSE, token.value, token.lineno, token.col_offset, token.error
+                        ), node
+                        continue
+
+                    separator_token = next((e for e in arguments if e.value == "AND"), None)
+                    args = []
+                    if separator_token is not None:
+                        args = arguments[: arguments.index(separator_token)]
+                        arguments = arguments[arguments.index(separator_token) :]
+                        has_separator = True
+                    else:
+                        if has_separator:
+                            args = arguments
+                            arguments = []
+
+                    async for e in self.generate_run_kw_tokens(
+                        namespace,
+                        builtin_library_doc,
+                        await namespace.find_keyword(unescape(token.value)) if is_not_variable_token(token) else None,
+                        RobotToken(ROBOT_KEYWORD_INNER, token.value, token.lineno, token.col_offset, token.error),
+                        args,
+                        node,
+                    ):
+                        yield e
+            elif kw_doc.is_run_keyword_if() and len(arguments) > 0:
+
+                async def generate_run_kw_if() -> AsyncGenerator[Tuple[Token, ast.AST], None]:
+                    nonlocal arguments
+
+                    yield arguments[0], node,
+                    arguments = arguments[1:]
+                    has_separator = False
+                    while arguments:
+                        async for b in skip_non_data_tokens():
+                            yield b
+
+                        if not arguments:
+                            break
+
+                        token = arguments[0]
+                        arguments = arguments[1:]
+
+                        if token.value in ["ELSE", "ELSE IF"]:
+                            yield RobotToken(
+                                RobotToken.ELSE, token.value, token.lineno, token.col_offset, token.error
+                            ), node
+
+                            if token.value == "ELSE IF":
+                                async for b in skip_non_data_tokens():
+                                    yield b
+
+                                if not arguments:
+                                    break
+
+                                yield arguments[0], node
+                                arguments = arguments[1:]
+                            continue
+
+                        inner_kw_doc = (
+                            await namespace.find_keyword(unescape(token.value))
+                            if is_not_variable_token(token)
+                            else None
+                        )
+
+                        if inner_kw_doc is not None and inner_kw_doc.is_run_keyword_if():
+                            yield RobotToken(
+                                ROBOT_KEYWORD_INNER, token.value, token.lineno, token.col_offset, token.error
+                            ), node
+
+                            arguments = arguments[1:]
+
+                            async for b in skip_non_data_tokens():
+                                yield b
+
+                            async for e in generate_run_kw_if():
+                                yield e
+
+                            continue
+
+                        separator_token = next((e for e in arguments if e.value in ["ELSE", "ELSE IF"]), None)
+                        args = []
+
+                        if separator_token is not None:
+                            args = arguments[: arguments.index(separator_token)]
+                            arguments = arguments[arguments.index(separator_token) :]
+                            has_separator = True
+                        else:
+                            if has_separator:
+                                args = arguments
+                                arguments = []
+
+                        async for e in self.generate_run_kw_tokens(
+                            namespace,
+                            builtin_library_doc,
+                            inner_kw_doc,
+                            RobotToken(ROBOT_KEYWORD_INNER, token.value, token.lineno, token.col_offset, token.error),
+                            args,
+                            node,
+                        ):
+                            yield e
+
+                async for e in generate_run_kw_if():
+                    yield e
+        else:
+
+            for a in arguments:
+                yield a, node
 
     async def collect(
         self,
         model: ast.AST,
         range: Optional[Range],
         namespace: Namespace,
-        finder: KeywordFinder,
         builtin_library_doc: Optional[LibraryDoc],
         cancel_token: CancelationToken,
     ) -> Union[SemanticTokens, SemanticTokensPartialResult, None]:
+
+        from robot.parsing.lexer.tokens import Token as RobotToken
+        from robot.parsing.model.statements import Fixture, KeywordCall
 
         data = []
         last_line = 0
         last_col = 0
 
-        def get_tokens() -> Generator[Tuple[Token, ast.AST], None, None]:
+        async def get_tokens() -> AsyncGenerator[Tuple[Token, ast.AST], None]:
             for node in iter_nodes(model):
                 if isinstance(node, HasTokens):
-                    for token in cast(HasTokens, node).tokens:
+                    if isinstance(node, (KeywordCall, Fixture)):
+                        kw_token = cast(
+                            Token,
+                            node.get_token(RobotToken.KEYWORD)
+                            if isinstance(node, KeywordCall)
+                            else node.get_token(RobotToken.NAME),
+                        )
+                        kw: Optional[str] = None
+
+                        for _, name in yield_owner_and_kw_names(kw_token.value):
+                            if name is not None:
+                                matcher = KeywordMatcher(name)
+                                if matcher in ALL_RUN_KEYWORDS_MATCHERS:
+                                    kw = name
+                        if kw:
+                            kw_doc = await namespace.find_keyword(kw_token.value)
+                            if kw_doc is not None and kw_doc.is_any_run_keyword():
+                                async for t in self.generate_run_kw_tokens(
+                                    namespace,
+                                    builtin_library_doc,
+                                    kw_doc,
+                                    kw_token,
+                                    node.tokens[node.tokens.index(kw_token) + 1 :],
+                                    node,
+                                ):
+                                    yield t
+                                continue
+
+                    for token in node.tokens:
                         yield token, node
 
-        for robot_token, robot_node in itertools.takewhile(
+        async for robot_token, robot_node in async_takewhile(
             lambda t: not cancel_token.throw_if_canceled() and (range is None or token_in_range(t[0], range)),
-            itertools.dropwhile(
+            async_dropwhile(
                 lambda t: not cancel_token.throw_if_canceled()
                 and range is not None
                 and not token_in_range(t[0], range),
@@ -352,10 +567,8 @@ class RobotSemanticTokenProtocolPart(RobotLanguageServerProtocolPart):
             ),
         ):
             cancel_token.throw_if_canceled()
-
-            async for token in self.generate_sem_tokens(
-                robot_token, robot_node, namespace, finder, builtin_library_doc
-            ):
+            self._logger.debug(repr(robot_token))
+            async for token in self.generate_sem_tokens(robot_token, robot_node, namespace, builtin_library_doc):
                 current_line = token.lineno - 1
 
                 data.append(current_line - last_line)
@@ -403,6 +616,7 @@ class RobotSemanticTokenProtocolPart(RobotLanguageServerProtocolPart):
                 ),
                 None,
             )
+            await namespace.get_library_doc()
 
             return await asyncio.get_running_loop().run_in_executor(
                 None,
@@ -411,7 +625,6 @@ class RobotSemanticTokenProtocolPart(RobotLanguageServerProtocolPart):
                     model,
                     range,
                     namespace,
-                    await namespace.create_finder(),
                     builtin_library_doc,
                     cancel_token,
                 ),