+++ CHANGELOG.md
@@ -7,7 +7,9 @@ All notable changes to the "robotcode" extension will be documented in this file
 ### added
 
 - some refactoring to speedup loading and parsing documents
-- semantic tokens now highlight builtin keywords
+- semantic tokens now highlight 
+  - builtin keywords
+  - run keywords
 - analysing run keywords now correctly unescape keywords
 
 ##  0.3.2